+++ packages/database/index.node.ts
@@ -86,8 +86,10 @@ export function registerDatabase(instance: FirebaseNamespace) {
         // getImmediate for FirebaseApp will always succeed
         const app = container.getProvider('app').getImmediate();
         const authProvider = container.getProvider('auth-internal');
+        const appCheckProvider = container.getProvider('app-check-internal');
+
         return new Database(
-          repoManagerDatabaseFromApp(app, authProvider, url),
+          repoManagerDatabaseFromApp(app, authProvider, appCheckProvider, url),
           app
         );
       },
@@ -123,8 +125,10 @@ try {
   // @firebase/app when used together with the js sdk. More detail:
   // https://github.com/firebase/firebase-js-sdk/issues/1696#issuecomment-501546596
   // eslint-disable-next-line import/no-extraneous-dependencies, @typescript-eslint/no-require-imports
-  const firebase = require('@firebase/app').default;
-  registerDatabase(firebase);
+  const firebase = require('@firebase/app').default; // Only present for v8, undefined for v9 (should skip).
+  if (firebase) {
+    registerDatabase(firebase);
+  }
 } catch (err) {
   // catch and ignore 'MODULE_NOT_FOUND' error in firebase-admin context
   // we can safely ignore this error because RTDB in firebase-admin works without @firebase/app