+++ packages/database/rollup.config.exp.js
@@ -71,7 +71,7 @@ const es5Builds = [
     input: 'exp/index.ts',
     output: [
       {
-        file: path.resolve('exp', expPkg.module),
+        file: path.resolve('exp', expPkg.esm5),
         format: 'es',
         sourcemap: true
       }
@@ -110,7 +110,7 @@ const es2017Builds = [
     input: 'exp/index.ts',
     output: [
       {
-        file: path.resolve('exp', expPkg.esm2017),
+        file: path.resolve('exp', expPkg.browser),
         format: 'es',
         sourcemap: true
       }