+++ packages/database/src/core/util/util.ts
@@ -26,8 +26,14 @@ import {
 } from '@firebase/util';
 
 import { SessionStorage } from '../storage/storage';
-import { QueryContext } from '../view/EventRegistration';
 
+// TODO: revert to import { QueryContext } from '../view/EventRegistration'; once the modular SDK goes GA
+/**
+ * This is part of a workaround for an issue in the no-modular '@firebase/database' where its typings
+ * reference types from `@firebase/app-exp`.
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type QueryContext = any;
 declare const window: Window;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare const Windows: any;