+++ packages/rules-unit-testing/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @firebase/rules-unit-testing
 
+## 1.2.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - firebase@8.4.3
+
+## 1.2.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - firebase@8.4.2
+
+## 1.2.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - firebase@8.4.1
+
 ## 1.2.8
 
 ### Patch Changes